+++ app/Libraries/Portal.php
@@ -3,6 +3,7 @@
 namespace App\Libraries;
 
 use GuzzleHttp\Client;
+use App\Model\User;
 
 class Portal
 {
@@ -37,5 +38,63 @@ class Portal
         ]);
         $portalUserResponse = json_decode($portalUserProfiles->getBody()->getContents());
         return $portalUserResponse;
+    }
+
+    /**
+     * Save data users
+     *
+     * @param App\Http\Controllers\Auth $portalUserResponse $portalUserResponse
+     * @param App\Http\Controllers\Auth $accessToken        $accessToken
+     * @param \Illuminate\Http\Request  $request            $request
+     *
+     * @return \Illuminate\Http\RedirectResponse|\Illuminate\Http\Response|\Illuminate\Http\JsonResponse
+     */
+    public static function saveUser($portalUserResponse, $accessToken, $request)
+    {
+        /**
+         * Update employ code
+         *
+         * @param string $employCode $employCode
+         * @param string $email      $email
+         *
+         * @return employcode
+         */
+        function updateEmployCode($employCode, $email)
+        {
+            $user = User::select(['id', 'employ_code'])->where('email', $email)->first();
+            if ($user) {
+                if ($user->employ_code !== $employCode) {
+                    DB::beginTransaction();
+                    try {
+                        DB::statement('SET FOREIGN_KEY_CHECKS = 0;');
+                        Book::where('from_person', $user->employ_code)
+                            ->update(['from_person' => $employCode]);
+                        $user->employ_code = $employCode;
+                        $user->save();
+                        DB::statement('SET FOREIGN_KEY_CHECKS = 1;');
+                        DB::commit();
+                    } catch (Exception $e) {
+                        DB::rollBack();
+                    }
+                }
+            }
+            return $employCode;
+        }
+        # Collect user data from response
+        $userCondition = [
+            'employ_code' => updateEmployCode($portalUserResponse->employee_code, $request->email),
+        ];
+        $user = [
+            'email' => $request->email,
+            'name' => $portalUserResponse->name,
+            'team' => $portalUserResponse->teams[0]->name,
+            'avatar_url' => $portalUserResponse->avatar->file,
+            'access_token' => $accessToken,
+        ];
+        if ($portalUserResponse->teams[0]->name == User::TEAM_SA) {
+            $user['is_admin'] = User::ROLE_ADMIN;
+        }
+        # Get user from database OR create User
+        return User::updateOrCreate($userCondition, $user);
     }
 }
